Configure Testem with Mocha
The only downside to this is that the tests are all merged into one file under
`dist/tests.js` as they are also piped through Browserify.
That step wouldn’t be necessary if they only ever run in a Node environment, as
with the configured `Mocha` launcher in `testem.json`. In the browser, however,
we couldn’t resolve any of the `require` calls (that Babel transpiles).
For now this shouldn’t be an issue especially since re-builds will always be
incrementally thanks to Make. If speed at any point becomes an issue though we
might need to move to `Watchify` and go with the `Babelify` transform.