Pass token to initialize callback
Basically all modules will need to make their own requests to the API and thus
add the appropriate `Authorization` header containing the token.

Instead of trying to offer a solution that conforms to all frameworks like to
offer a proxy mechanism the all modules have to use, passing the token down
and letting the module handle this seems most sensible.

The last thing will be to decide on how to handle `401` responses from the API
and how to expose the `unauthorize` interface.