Fix regression in serializing url-parse objects
When `delete`ing from the `query` hash, the object’s `href` is not updated
properly as it does not know about property changes unless they’re made through
`.set`, which in the case of a parsed `query` hash does not work (at least
`set`ting it to `undefined` or `''` didn’t do the trick).

By passing the object directly to `history.replaceState` it is stringified,
which then properly serialized its current properties, so `delete` works as
expected again. This is also nicer since the `location` module does not need to
know anything about the parameter and instead just passes it on.