Make test target depend on app build files as well
Otherwise, the test bundle is not re-compiled when you touch only app code.
Especially when running `make test` in a separate process you won’t see
automatic test re-runs, which can be quite misleading.

There is still some weird issue with Chokidar going on that sometimes triggers
twice when changing a file but it does not impact the test runs or
compilation/bundling.